# openapi.yaml
openapi: 3.0.0
info:
  title: Rust Database API
  description: |
    RESTful API for a thread-safe, concurrent Rust database engine.
    Supports CRUD operations on tables, columns, and rows with multiple data types.
  version: 1.0.0
  contact:
    name: DB API Support
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:8080
    description: Local development server
  - url: http://localhost:8080
    description: Localhost
...
paths:
...
  /api/tables:
    post:
      tags:
        - Tables
      summary: Create a new table
      description: Create a new table with the specified name
      operationId: createTable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTableRequest'
      responses:
        '201':
          description: Table created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTableResponse'
        '500':
          description: Failed to create table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Tables
      summary: List all tables
      description: Get a list of all tables in the database
      operationId: listTables
      responses:
        '200':
          description: List of tables retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTablesResponse'
        '503':
          description: Database unavailable (lock poisoned)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tables/by-name/{name}:
    get:
      tags:
        - Tables
      summary: Get table ID by name
      description: Retrieve the table ID for a table by its name
      operationId: getTableByName
      parameters:
        - name: name
          in: path
          required: true
          description: Table name
          schema:
            type: string
      responses:
        '200':
          description: Table found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTableByNameResponse'
        '404':
          description: Table not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tables/{table_id}:
    delete:
      tags:
        - Tables
      summary: Drop a table
      description: Delete a table and all its data
      operationId: dropTable
      parameters:
        - name: table_id
          in: path
          required: true
          description: Table ID
          schema:
            type: integer
            format: uint32
      responses:
        '200':
          description: Table dropped successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DropTableResponse'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tables/{table_id}/schema:
    get:
      tags:
        - Tables
      summary: Get table schema
      description: Retrieve the schema (columns and types) of a table
      operationId: getTableSchema
      parameters:
        - name: table_id
          in: path
          required: true
          description: Table ID
          schema:
            type: integer
            format: uint32
      responses:
        '200':
          description: Schema retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSchemaResponse'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tables/{table_id}/size:
    get:
      tags:
        - Tables
      summary: Get table row count
      description: Get the number of rows in a table
      operationId: getTableSize
      parameters:
        - name: table_id
          in: path
          required: true
          description: Table ID
          schema:
            type: integer
            format: uint32
      responses:
        '200':
          description: Row count retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSizeResponse'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tables/{table_id}/columns:
    post:
      tags:
        - Columns
      summary: Create a new column
      description: Add a new column to a table
      operationId: createColumn
      parameters:
        - name: table_id
          in: path
          required: true
          description: Table ID
          schema:
            type: integer
            format: uint32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateColumnRequest'
      responses:
        '201':
          description: Column created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateColumnResponse'
        '500':
          description: Failed to create column
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tables/{table_id}/columns/{column_id}:
    delete:
      tags:
        - Columns
      summary: Drop a column
      description: Delete a column from a table
      operationId: dropColumn
      parameters:
        - name: table_id
          in: path
          required: true
          description: Table ID
          schema:
            type: integer
            format: uint32
        - name: column_id
          in: path
          required: true
          description: Column ID
          schema:
            type: integer
            format: uint32
      responses:
        '200':
          description: Column dropped successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DropColumnResponse'
        '500':
          description: Failed to drop column
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tables/{table_id}/columns/{column_id}/index:
    post:
      tags:
        - Columns
      summary: Create an index
      description: Create an index on a column for faster lookups. Automatically created for columns with Unique or PrimaryKey constraints.
      operationId: createIndex
      parameters:
        - name: table_id
          in: path
          required: true
          description: Table ID
          schema:
            type: integer
            format: uint32
        - name: column_id
          in: path
          required: true
          description: Column ID
          schema:
            type: integer
            format: uint32
      responses:
        '201':
          description: Index created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateIndexResponse'
        '500':
          description: Failed to create index
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Columns
      summary: Drop an index
      description: Remove an index from a column
      operationId: dropIndex
      parameters:
        - name: table_id
          in: path
          required: true
          description: Table ID
          schema:
            type: integer
            format: uint32
        - name: column_id
          in: path
          required: true
          description: Column ID
          schema:
            type: integer
            format: uint32
      responses:
        '200':
          description: Index dropped successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DropIndexResponse'
        '500':
          description: Failed to drop index
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/tables/{table_id}/rows/{row_id}:
    get:
      tags:
        - Rows
      summary: Get a specific row
      description: Retrieve a row by its ID
      operationId: getRow
      parameters:
        - name: table_id
          in: path
          required: true
          description: Table ID
          schema:
            type: integer
            format: uint64
        - name: row_id
          in: path
          required: true
          description: Row ID
          schema:
            type: integer
            format: uint64
      responses:
        '200':
          description: Row retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRowResponse'
        '404':
          description: Row not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tables/{table_id}/search:
    post:
      tags:
        - Rows
      summary: Search rows
      description: Search and filter rows with optional projection and sorting. Allows complex queries with multiple search criteria, column selection (projection), and result ordering.
      operationId: searchRows
      parameters:
        - name: table_id
          in: path
          required: true
          description: Table ID
          schema:
            type: integer
            format: uint32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRowsRequest'
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchRowsResponse'
        '400':
          description: Invalid search criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to execute search
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    CreateTableRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Table name
          example: users
    CreateTableResponse:
      type: object
      properties:
        table_id:
          type: integer
          format: uint32
          example: 1
    ListTablesResponse:
      type: object
      properties:
        tables:
          type: array
          items:
            type: object
            properties:
              table_id:
                type: integer
                format: uint32
              table_name:
                type: string
    GetTableByNameResponse:
      type: object
      properties:
        table_id:
          type: integer
          format: uint32
    DropTableResponse:
      type: object
      properties:
        success:
          type: boolean
    GetSchemaResponse:
      type: object
      properties:
        columns:
          type: array
          items:
            $ref: '#/components/schemas/ColumnInfo'
...
    GetSizeResponse:
      type: object
      properties:
        row_count:
          type: integer
          format: uint32
    CreateColumnRequest:
      type: object
      required:
        - column_name
        - data_type
      properties:
        column_name:
          type: string
          example: email
        data_type:
          type: string
          enum:
            - IntegerU8
            - IntegerI128
            - String
            - Boolean
          example: String
...
    DropColumnResponse:
      type: object
      properties:
        success:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Internal server error"
...
  securitySchemes: {}
security: []
...